requirements.txt
pyyaml==6.0.2
openpyxl==3.1.5
numpy==2.2.3
matplotlib==3.10.1
pandas==2.2.3
ipython==9.0.2